rework the sanitizer
Until we can render HTML in table cells, the current approach -- strip HTML
tags, then massage the output -- is a compromise solution. One variable is
whether we use the sanitizer like this:

var sanitizer = bluemonday.StrictPolicy().AddSpaceWhenStrippingTag(true)

Or without that option. This commit removes AddSpaceWhenStrippingTag(true)
which sometimes allows text to run together when it shouldn't, but handles
URLs (into which Mastodon inserts span tags for truncation) better.

It's an awkward compromise.